Feat: 운동 기록 검색 및 삭제 기능 추가

+++ hwpSpring/src/main/java/com/hwpBackend/hwpSpring/controller/ExerciseRecordController.java
@@ -1,19 +1,25 @@
 package com.hwpBackend.hwpSpring.controller;
 
 import com.hwpBackend.hwpSpring.entity.ExerciseRecord;
+import com.hwpBackend.hwpSpring.entity.Member;
 import com.hwpBackend.hwpSpring.exception.InfoOrRecordNotFoundException;
 import com.hwpBackend.hwpSpring.repository.ExerciseRecordRepository;
+import com.hwpBackend.hwpSpring.util.SecurityUtil;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @RestController
+@RequestMapping("/exerciseRecords")
 public class ExerciseRecordController {
-    private ExerciseRecordRepository repository;
+    private final ExerciseRecordRepository repository;
 
     public ExerciseRecordController(ExerciseRecordRepository repository) {
         this.repository = repository;
@@ -22,33 +28,58 @@ public class ExerciseRecordController {
     ;
 
 
-    @GetMapping("/exerciseRecords")
-    public List<ExerciseRecord> retrieveAllExerciseRecord() {
+//    @GetMapping("/exerciseRecords")
+//    public List<ExerciseRecord> retrieveAllExerciseRecord() {
+//        return repository.findAll();
+//    }
+
+    @GetMapping("/")
+    public List<ExerciseRecord> retrieveAllExerciseRecord(){
         return repository.findAll();
     }
 
-    @GetMapping("/exerciseRecords/{id}")
-    public Optional<ExerciseRecord> retrieveExerciseRecord(@PathVariable(value = "id") Integer id) {
-        Optional<ExerciseRecord> exerciseRecord = repository.findById(id);
+    @GetMapping("/{id}")
+    public ResponseEntity<?> retrieveExerciseRecord(@PathVariable(value = "id") String id) {
+        String currentUser = SecurityUtil.getCurrentUsername();
 
-        if (exerciseRecord.isEmpty()) throw new InfoOrRecordNotFoundException("id:" + id.toString());
-        return exerciseRecord;
+        if(!currentUser.equals(id)) {
+            throw new AccessDeniedException("본인의 정보가 아닙니다.");
+        }
+        List<ExerciseRecord> savedExerciseRecord = repository.findByMember_Username(id);
+
+        if (savedExerciseRecord.isEmpty()) throw new InfoOrRecordNotFoundException("id:" + id.toString());
+        return new ResponseEntity<>(savedExerciseRecord, HttpStatus.OK);
     }
 
-    @PostMapping("/exerciseRecords")
+
+    @PostMapping("/add")
     public ResponseEntity<ExerciseRecord> createExerciseRecord(@RequestBody ExerciseRecord exerciseRecord) {
         ExerciseRecord savedExerciseRecord = repository.save(exerciseRecord);
 
         URI location = ServletUriComponentsBuilder.fromCurrentRequest()
                 .path("/{id}")
-                .buildAndExpand(savedExerciseRecord.getRecordID())
+                .buildAndExpand(savedExerciseRecord.getRecordId())
                 .toUri();
 
         return ResponseEntity.created(location).build();
     }
 
-    @DeleteMapping("/exerciseRecords/{id}")
+    @DeleteMapping("/{id}/delete")
     public void deleteExerciseRecord(@PathVariable(value = "id") Integer id) { // String인 경우 반드시 value값을 지정해줄 것
-        repository.deleteById(id);
+        String currentUser = SecurityUtil.getCurrentUsername();
+
+        // findById 메서드의 결과가 null이면 해당 id에 해당하는 정보가 없다는 뜻
+        ExerciseRecord record = repository.findById(id).orElse(null);
+
+        if(record != null && currentUser.equals(Objects.requireNonNull(record).getMember().getUsername())){
+            repository.deleteById(id);
+        }
+        else if (record == null){
+            throw new InfoOrRecordNotFoundException("해당하는 정보를 찾을 수 없습니다.");
+        }
+        else{
+            // 본인의 정보가 아님
+            throw new AccessDeniedException("본인의 정보가 아닙니다.");
+        }
     }
 }
